% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebna.R
\docType{data}
\name{ebna}
\alias{ebna}
\title{HLA allele summary statistics for EBNA}
\format{
A dataframe with 168 rows and 4 variables:
\describe{
\item{Name}{chr Name of HLA allele in IMGT-HLA format.}
\item{beta}{num Beta of HLA allele association.}
\item{se}{numStandard error of HLA allele association.}
\item{N}{int Sample size of HLA association.}
}
}
\usage{
ebna
}
\description{
HLA allele summary statistics for EBNA
}
\keyword{datasets}
