% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_coloc.R
\name{hla_coloc}
\alias{hla_coloc}
\title{Perform HLA colocalization}
\usage{
hla_coloc(
  pheno1,
  pheno1R,
  is_cohort_ld_pheno1 = FALSE,
  pheno2,
  pheno2R,
  is_cohort_ld_pheno2 = FALSE,
  max_iter_susieR = 100,
  plot_susie = TRUE,
  negative_threshold = 0.001,
  susie_L = 10,
  n_min_alleles = 10
)
}
\arguments{
\item{pheno1}{Dataframe of HLA allele associations for phenotype 1. Needs to contain the following four columns:
\itemize{
\item \code{Name}: the name of the HLA allele in the IMGT-HLA format.
\item \code{beta}: the beta (effect sizes) of the HLA allele association.
\item \code{se}: the standard error.
\item \code{N}: the sample size.
}}

\item{pheno1R}{A dataframe or matrix of correlation coefficient (R) or the alleles in pheno1. The rows and columns need to be in the same order as the alleles in pheno1.}

\item{is_cohort_ld_pheno1}{Whether the LD matrix for the pheno1 cohort is from the same cohort (TRUE) or from an external reference (FALSE).}

\item{pheno2}{Dataframe of HLA allele associations for phenotype 2. Needs to be in the same format as pheno1, including rows in same order.}

\item{pheno2R}{A dataframe or matrix of correlation coefficient (R) or the alleles in pheno2. The rows and columns need to be in the same order as the alleles in pheno2.}

\item{is_cohort_ld_pheno2}{Whether the LD matrix for the pheno2 cohort is from the same cohort (TRUE) or from an external reference (FALSE).}

\item{max_iter_susieR}{Maximum number of iterations for susieR (default=100).}

\item{plot_susie}{Whether to plot the results (TRUE/FALSE).}

\item{negative_threshold}{Minimum susieR posterior inclusion probability needed for both phenotypoes in order to check for colocalization using stanR (default=0.001).}

\item{susie_L}{Maximum number of alleles with non-zero effect in the susieR model (default=10).}

\item{n_min_alleles}{Minimum number of alleles required at a gene in order to attempt HLA colocalization at that gene. Genes with less than this threshold will be excluded from the analyses (default=10).}
}
\value{
List with a dataframe containing the colocalization result, and a plot of the results (optional).
}
\description{
Perform HLA colocalization
}
