% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\docType{data}
\name{ms}
\alias{ms}
\title{HLA allele summary statistics for MS}
\format{
A dataframe with 168 rows and 4 variables:
\describe{
\item{Name}{chr Name of HLA allele in IMGT-HLA format.}
\item{beta}{num Beta of HLA allele association.}
\item{se}{numStandard error of HLA allele association.}
\item{N}{int Sample size of HLA association.}
}
}
\usage{
ms
}
\description{
HLA allele summary statistics for MS
}
\keyword{datasets}
