% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_ebna.R
\docType{data}
\name{r_ebna}
\alias{r_ebna}
\title{HLA allele LD coefficient R for EBNA}
\format{
A dataframe with 168 rows and 168 variables, for each HLA allele (in same order as ebna):
}
\usage{
r_ebna
}
\description{
HLA allele LD coefficient R for EBNA
}
\keyword{datasets}
