% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_ms.R
\docType{data}
\name{r_ms}
\alias{r_ms}
\title{HLA allele LD coefficient R for MS}
\format{
A dataframe with 168 rows and 168 variables, for each HLA allele (in same order as ms):
}
\usage{
r_ms
}
\description{
HLA allele LD coefficient R for MS
}
\keyword{datasets}
